favs view: replace favs ids sub with effect function

+++ src/app/components/favourites-view/favourites-view.component.ts
@@ -1,8 +1,7 @@
-import { HitIds } from '../../models/study';
 import { FavouritesService } from '../../services/favourites.service';
 import { FavsStudiesStore } from '../../stores/favs-studies.store';
 import { StudyCardComponent } from '../study-card/study-card.component';
-import { Component, inject, OnDestroy, OnInit } from '@angular/core';
+import { Component, effect, inject, untracked } from '@angular/core';
 import { patchState } from '@ngrx/signals';
 import { Subscription } from 'rxjs';
 
@@ -14,30 +13,59 @@ import { Subscription } from 'rxjs';
   templateUrl: './favourites-view.component.html',
   styleUrl: './favourites-view.component.scss',
 })
-export class FavouritesViewComponent implements OnInit, OnDestroy {
+export class FavouritesViewComponent {
   protected favouritesService = inject(FavouritesService);
   readonly favsStudyStore = inject(FavsStudiesStore);
 
   favIdsSub!: Subscription;
 
-  ngOnInit(): void {
+  constructor() {
     this.favouritesService.resetDeletedFavourite();
-    this.favIdsSub = this.favouritesService.favourites$.subscribe(
-      (favs: HitIds) => {
-        if (this.favouritesService.deletedFavourite() === null) {
-          this.favsStudyStore.loadFavouriteStudies(favs);
-        } else {
+
+    effect(
+      () => {
+        this.updateFavsStudies();
+      },
+      { allowSignalWrites: true }
+    );
+  }
+
+  updateFavsStudies(): void {
+    if (this.favouritesService.deletedFavourite() === null) {
+      this.favsStudyStore.loadFavouriteStudies(
+        this.favouritesService.favourites()
+      );
+    } else {
+      if (this.favouritesService.favourites().length) {
+        untracked(() => {
           const newStudies = this.favsStudyStore
             .favouriteStudies()
             .filter(
               (study) => study.id !== this.favouritesService.deletedFavourite()
             );
           patchState(this.favsStudyStore, { favouriteStudies: newStudies });
-        }
+        });
+      } else {
+        patchState(this.favsStudyStore, { favouriteStudies: [] });
       }
-    );
+    }
   }
-  ngOnDestroy(): void {
-    this.favIdsSub.unsubscribe();
-  }
+
+  // ngOnInit(): void {
+  // this.favouritesService.resetDeletedFavourite();
+  // this.favIdsSub = this.favouritesService.favourites$.subscribe(
+  //   (favs: HitIds) => {
+  // if (this.favouritesService.deletedFavourite() === null) {
+  //   this.favsStudyStore.loadFavouriteStudies(favs);
+  // } else {
+  //   const newStudies = this.favsStudyStore
+  //     .favouriteStudies()
+  //     .filter(
+  //       (study) => study.id !== this.favouritesService.deletedFavourite()
+  //     );
+  //   patchState(this.favsStudyStore, { favouriteStudies: newStudies });
+  // }
+  // }
+  // );
+  // }
 }
